Refactor budgets controller and specs with log_in_as helper
