+++ app/views/budgets/index.html.erb
@@ -1,5 +1,4 @@
-<div><%= t(".title") %></div>
-
+<h1><%= t(".title") %></h1>
 <ul>
   <% @budgets.each do |budget| %>
     <%= render "detail", budget: budget %>
