+++ app/views/budgets/show.html.erb
@@ -1,5 +1,4 @@
-<div><%= t(".title") %></div>
-
+<h1><%= t(".title") %></h1>
 <ul>
   <%= render "detail" %>
 </ul>
