+++ spec/features/sessions/user_creates_new_session_spec.rb
@@ -3,29 +3,21 @@ require "rails_helper"
 feature "user creates new session" do
   scenario "successfully" do
     user = create(:user)
-
-    visit new_session_path
-    fill_in "session_email", with: user.email
-    fill_in "session_password", with: user.password
-    click_on I18n.t("sessions.new.submit")
+    log_in_as user
 
     expect(page).to have_text(I18n.t("users.show.welcome", name: user.name))
-    expect(page).to have_text(I18n.t("sessions.create.notice"))
+    expect(page).to have_text(I18n.t("sessions.create.flash.notice"))
   end
 
   scenario "logs out" do
     user = create(:user)
-
-    visit new_session_path
-    fill_in "session_email", with: user.email
-    fill_in "session_password", with: user.password
-    click_on I18n.t("sessions.new.submit")
+    log_in_as user
 
     click_on I18n.t("sessions.destroy.link")
 
     expect(page).to have_text(I18n.t("home.show.welcome"))
     expect(page).to have_text(
-      I18n.t("sessions.destroy.notice", resource_name: "Session")
+      I18n.t("sessions.destroy.flash.notice", resource_name: "Session")
     )
   end
 end
