+++ spec/spec_helper.rb
@@ -11,7 +11,6 @@ end
 
 require "webmock/rspec"
 require "timecop"
-require "helpers/helpers"
 
 # http://rubydoc.info/gems/rspec-core/RSpec/Core/Configuration
 RSpec.configure do |config|
@@ -26,8 +25,6 @@ RSpec.configure do |config|
 
   config.example_status_persistence_file_path = "tmp/rspec_examples.txt"
   config.order = :random
-
-  config.include Helpers
 end
 
 WebMock.disable_net_connect!(allow_localhost: true)
